ci: WASM/BDD test for MacOS
