+++ test/aries-js-worker/package.json
@@ -5,8 +5,8 @@
   "description": "Tests",
   "main": "index.js",
   "scripts": {
-    "test": "bash scripts/setup_assets.sh && karma start",
-    "test-local": "bash scripts/setup_assets.sh && karma start --browsers=Chrome_without_security --single-run=false"
+    "test": "bash scripts/setup_assets.sh && node init.js && karma start",
+    "test-local": "bash scripts/setup_assets.sh && node init.js && karma start --browsers=Chrome_without_security --single-run=false"
   },
   "author": "hyperledger/aries",
   "license": "Apache-2.0",
